calculating the payment amount and time ti give unlock/activation codes
