+++ shared/getCurrentDate.js
@@ -1,0 +1,4 @@
+module.exports = function(){
+    var date = new Date();
+    return date;
+};
