+++ shs/helper-functions/clientTestData.js
@@ -1,0 +1,1 @@
+module.exports = {};
